+++ akshare/fund/fund_portfolio_em.py
@@ -67,6 +67,55 @@ def fund_portfolio_hold_em(symbol: str = "162411", date: str = "2020") -> pd.Dat
     return big_df
 
 
+def fund_portfolio_bond_hold_em(symbol: str = "000001", date: str = "2021") -> pd.DataFrame:
+    """
+    天天基金网-基金档案-投资组合-债券持仓
+    http://fundf10.eastmoney.com/ccmx1_000001.html
+    :param symbol: 基金代码
+    :type symbol: str
+    :param date: 查询年份
+    :type date: str
+    :return: 债券持仓
+    :rtype: pandas.DataFrame
+    """
+    url = "http://fundf10.eastmoney.com/FundArchivesDatas.aspx"
+    params = {
+        "type": "zqcc",
+        "code": symbol,
+        "year": date,
+        "rt": "0.913877030254846",
+    }
+    r = requests.get(url, params=params)
+    data_text = r.text
+    data_json = demjson.decode(data_text[data_text.find("{") : -1])
+    soup = BeautifulSoup(data_json["content"], "lxml")
+    item_label = [
+        item.text.split("\xa0\xa0")[1]
+        for item in soup.find_all("h4", attrs={"class": "t"})
+    ]
+    big_df = pd.DataFrame()
+    for item in range(len(item_label)):
+        temp_df = pd.read_html(data_json["content"], converters={"债券代码": str})[item]
+        temp_df["占净值比例"] = temp_df["占净值比例"].str.split("%", expand=True).iloc[:, 0]
+        temp_df.rename(columns={"持仓市值（万元）": "持仓市值"}, inplace=True)
+        temp_df["季度"] = item_label[item]
+        temp_df = temp_df[
+            [
+                "序号",
+                "债券代码",
+                "债券名称",
+                "占净值比例",
+                "持仓市值",
+                "季度",
+            ]
+        ]
+        big_df = big_df.append(temp_df, ignore_index=True)
+    big_df["占净值比例"] = pd.to_numeric(big_df["占净值比例"], errors="coerce")
+    big_df["持仓市值"] = pd.to_numeric(big_df["持仓市值"], errors="coerce")
+    big_df["序号"] = range(1, len(big_df) + 1)
+    return big_df
+
+
 def fund_portfolio_change_em(
     symbol: str = "003567", indicator: str = "累计买入", date: str = "2020"
 ) -> pd.DataFrame:
@@ -142,6 +191,9 @@ def fund_portfolio_change_em(
 if __name__ == "__main__":
     fund_portfolio_hold_em_df = fund_portfolio_hold_em(symbol="162411", date="2020")
     print(fund_portfolio_hold_em_df)
+
+    fund_portfolio_bond_hold_em_df = fund_portfolio_bond_hold_em(symbol="000001", date="2021")
+    print(fund_portfolio_bond_hold_em_df)
 
     fund_portfolio_change_em_df = fund_portfolio_change_em(
         symbol="003567", indicator="累计买入", date="2020"